feat: allow sequence classifier handle tensors from ListField (and refactor)
